Add .gitignore file
This commit introduces a .gitignore file to the project. The .gitignore file is used to tell Git which files or directories to ignore in a project. In this case, it's set to ignore Python cache files (__pycache__) and compiled Python files (*.pyc). This is useful to keep the repository clean and avoid committing unnecessary files.